Since VirtualEarth doesn't understand bounds at all, this function didn't work. This is not an ideal situation, becauuse it's not really 100% correct for a mercator projection, but it does cause things to work.
It seems that for some reason, the EventPane isn't actually catching all the 
move events -- this seems to be what is causing the slow dragging. I'm not 
sure of the reason for this, but maybe Erik or someone else will be able to 
offer hints. I could also just be wrong ;) I just know it's slow, really.