Firefox SVG does not support drawing points with values more than something between 10,000 and 40,000. (It's hard to tell, since it crashes the browser when you try.) The resulting behavior is that the browser crashes. This wraps these values into Pixel regions, and doesn't draw features which fall outside of them. (This sucks, but the result is the browser not crashing.) This closes #669, and should be improved as described in #670 for 2.5.
If for some reason you know that the browser you're on can support larger 
maxPixel values, you can adjust the maxPixel private variable on the 
renderer. I believe that Opera and WebKit nightlies will work fine with 
a larger maxPixel value. Changing this is encouraged *only* if you 
are developer working on the code, because it can totally crash browsers 
in a production environment.