Add support for a Canvas renderer. This renderer will allow for some new capabilties in OpenLayers Vector drawing, including:  * Vector support for iPhone, Safari 2.x series browsers  * Improved performance of dragging the map with a large number of    geometries.
The Vector layer default renderer order is now SVG, VML, Canvas, so browsers
with Canvas support and no SVG or VML will be able to draw vectors.

The Canvas layer has a number of limitations: getFeatureFromEvent is much
slower than the other types of layer, and any change to any feature requires a
redraw of the entire canvas. Because Canvas is typically a pretty fast 
drawing implementation, the latter is less problematic than it might otherwise
be.

r=pagameba,fred, with glances from a couple other people.
(Closes #1512)